% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.r
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(name)
}
\arguments{
\item{name}{a name input by user, like 'Dave'}
}
\value{
A string like 'Hello, Dave!'
}
\description{
output Hello, <name> whereas <name> is the user input
}
\examples{
name = 'Dave'
greeting_str = hello(name='Dave')
}
